Add files via upload
A Restful Service using PHP with Mysql. Included is a MySQL table extract from the famous Northwind Database employee table.
In case you need to use or test it. You must import the Employee.sql into your local PHPMysqlAdmin to create the Table, Make changes to your table and database attributes in both include/db_connect.php, and also in class/rest_api.php in the following lines.

// db_connect.php
$domainhost = "localhost";
$username = "root";
$password = "";
$dbname = "northwind";

and also in the following file
// rest_api.php
class Rest_api{
    private $domainhost  = 'localhost';
    private $username  = 'root';
    private $password   = "";
    private $database  = "northwind";      
    private $employeeTable = 'employee';

For the program to work for you.